feat: Add PHP code to list products from the database
The code changes include adding PHP code to the `listaProdutos.php` file to retrieve and display a list of products from the database. The code uses PDO to connect to the database, prepares and executes an SQL query to select all products from the `produtos` table, and then loops through the results to display each product in an HTML table.